rest_grpc_lab/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
grpcio==1.74.0
grpcio-tools==1.74.0
httpx==0.28.1
psycopg2-binary==2.9.7
